+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # LIFX Changelog
 
 ## [Initial Version] - 2022-08-12
+- Added `toggle light`command
+- Added `set light` command
+- added `view lights` command